+++ threedi_modelchecker/config.py
@@ -1019,19 +1019,22 @@ CHECKS += [
         "connection_node.exchange_type is CONNECTED. In the future, this will lead to an error.",
     ),
     QueryCheck(
-        level=CheckLevel.WARNING,
+        level=CheckLevel.ERROR,
         error_code=107,
         column=models.ConnectionNode.exchange_level,
-        filters=CONDITIONS["has_no_dem"]
-        .filter(models.ModelSettings.manhole_aboveground_storage_area > 0)
-        .exists(),
+        filters=or_(
+            CONDITIONS["has_dem"].exists(),
+            Query(models.ModelSettings)
+            .filter(models.ModelSettings.manhole_aboveground_storage_area > 0)
+            .exists(),  # to avoid SQLAlchemy complaints about cartesian products
+        ),
         invalid=Query(models.ConnectionNode).filter(
             models.ConnectionNode.exchange_type.in_(
                 [constants.CalculationTypeNode.CONNECTED]
             ),
             models.ConnectionNode.exchange_level == None,
         ),
-        message="connection_node.exchange_level cannot be null when using sub-basins (model_settings.manhole_aboveground_storage_area > 0) and no DEM is supplied.",
+        message="connection_node.exchange_level cannot be null when using sub-basins (model_settings.manhole_aboveground_storage_area > 0) and/or a DEM is supplied.",
     ),
 ]
 CHECKS += [